Definitions of components and temp implementations
Moved the old code to main_old, the code will be moved into listener and measurement.
